% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addCorNoise}
\alias{addCorNoise}
\title{Adds some multivariate gaussian noise to a correlation matrix.
The noise matrix is made symmetric so that the covariance matrix is
still Hermitian (i.e., real eigenvalues, etc.)}
\usage{
addCorNoise(mat, config)
}
\value{
a matrix
}
\description{
Adds some multivariate gaussian noise to a correlation matrix.
The noise matrix is made symmetric so that the covariance matrix is
still Hermitian (i.e., real eigenvalues, etc.)
}

