% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_based_tests.R
\name{ccaTest}
\alias{ccaTest}
\title{Applies Manuel Ferreira's canoncical correlation test}
\usage{
ccaTest(gene, pheno)
}
\description{
Applies Manuel Ferreira's canoncical correlation test
}

