% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.gwasresult}
\alias{plot.gwasresult}
\title{Plot simulated GWAS result data}
\usage{
\method{plot}{gwasresult}(x, type = "manhattan")
}
\arguments{
\item{x}{the result of a simulated GWAS}

\item{type}{specificies manhattan vs qqplot}
}
\description{
Plot simulated GWAS result data
}

