% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{setConfiguration}
\alias{setConfiguration}
\title{sets the configuration for the simulations}
\usage{
setConfiguration(LD, N_BLOCKS, N_MARKERS)
}
\arguments{
\item{LD}{can be set as low, medium, or high}

\item{N_BLOCKS}{is the number of LD blocks generated}

\item{N_MARKERS}{is the number of SNPs per LD block}
}
\description{
sets the configuration for the simulations
}
\details{
These parameters are passed on to \link{setConfiguration_}
which creates the list configruation object. By default
one causal SNP is used with an effect size of .01 Rsquared.
The phenotype is assumed to be Gaussian.

This function is intended to provide a simplied interface
to the setConfiguration_ function.
}

