% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setConfiguration_}
\alias{setConfiguration_}
\title{called by \link{setConfiguration}}
\usage{
setConfiguration_(N_MARKERS, LD, simple_interface, sigma, N_COV, ALLELEFRQ,
  N_STRANDS, N_ALLELES = N_MARKERS * 2, N_BLOCKS, BLOCK_COR,
  EFFECT_SIZE = 0.01, N_CAUSAL = 5, N_CAUSAL_PER_BLOCK = 1,
  PHENO_SD = 3, PHENO_DIST = "gaussian", COR_NOISE_VAR = 0)
}
\value{
Returns a list
}
\description{
called by \link{setConfiguration}
}
\details{
advanced users are recommended to simply view the source code to view all
available options by typing in setConfiguration_ into an interactive R console
}

