% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simBlock}
\alias{simBlock}
\title{Simulate an LD block}
\usage{
simBlock(config)
}
\arguments{
\item{config}{An object created by \link{setConfiguration}}
}
\details{
Generates two chromosomes seperately, and sums them.
Each is generated using a multivariate binomial random variable generator.
See bindata::rmvbin for more details.
}

