% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simBlockR}
\alias{simBlockR}
\title{Find the correlation between two blocks}
\usage{
simBlockR(block1, block2)
}
\arguments{
\item{block1}{A LD block}

\item{block2}{A LD block}
}
\value{
the correlation between two LD blocks
}
\details{
canonical correlation is used to determine
 the correlation between two matrices of the same dimension
}

