% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simBlockSet}
\alias{simBlockSet}
\title{Simulated genotype data}
\usage{
simBlockSet(config)
}
\arguments{
\item{config}{A list cretaed by \link{setConfiguration}}
}
\value{
a matrix of several adjacent LD blocks of type genosim
}
\description{
generates a matrix of several LD blocks combined together
}
\details{
The LD blocks are generated separately and are combined
in a greedy fashion. The process is generally as follows:
1) generate block_1 using the \link{simBlock} function
2) generate block_2
3) compare block_1 to block_2 using canonical correlation.
4) Accept block_2 as the next block in the sequence of LD blocks
   if the between block canonical correlation exceeds the user
   defined BLOCK_COR parameter.
}

