% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{simCov}
\alias{simCov}
\title{Simulates a covariance matrix for an LD block}
\usage{
simCov(config, limitZ = 1)

simCov(config, limitZ = 1)
}
\arguments{
\item{config}{An object created by \link{setConfiguration}}

\item{config}{An object created by \link{setConfiguration}}
}
\value{
a matrix representing the LD of the block
}
\description{
Simulates a covariance matrix for an LD block
}
\details{
Covariance matrix is drawn from Wishart distribution
using the \link{rWishart} function. The most important parameter
to simulate a covariance matrix is the "population" covariance matrix
specified in the configuration object.
This function also ensures that the resulting covariance matrix
is positive definite using the nearPD function from the matrix package.
}

