% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simGWAS}
\alias{simGWAS}
\title{Get simulated GWAS results}
\usage{
simGWAS(gene, pheno, config, parallel = F, cl)
}
\arguments{
\item{gene}{the result of a call to \link{simBlockSet}}

\item{pheno}{the result of a call to \link{simPhenotype}}

\item{config}{An object created by \link{setConfiguration}}

\item{parallel}{A logical to determine whether or not the analysis should
be run in parallel.}

\item{cl}{a parallel cluster if running the analysis in parallel}
}
\value{
an object of class gwasresult
}

