% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate_phenotype.R
\name{simPhenotype}
\alias{simPhenotype}
\title{Add together two numbers.}
\usage{
simPhenotype(config, gene)
}
\arguments{
\item{config}{A configuration object.}

\item{gene}{A genosim object.}
}
\value{
An object of class phenosim. Includes simulated phenotype,
causal blocks, causal markers, and structure
}
\description{
Add together two numbers.
}

