% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snpTest}
\alias{snpTest}
\title{Run association test between SNP and phenotype}
\usage{
snpTest(snpName, dat, dist)
}
\arguments{
\item{snpName}{the name of the snp to run the test on}

\item{dat}{the dataframe with the genotype and phenotype attached}

\item{dist}{the probability distribution of the phenotype
This function takes the genotype and phenotype information and runs
the association test between the given SNP and the phenotype.
The association test is either linear or logistic regression as determined
by the distribution of the phenotype.}
}
\value{
the pvalue
}

