% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenotype.R
\name{simPhenotype}
\alias{simPhenotype}
\title{Simulate phenotype data}
\usage{
simPhenotype(config, gene, verbose = F)
}
\arguments{
\item{config}{A configuration object.}

\item{gene}{A genosim object.}
}
\value{
An object of class phenosim. Includes simulated phenotype,
causal blocks, causal markers, and structure
}

